+++ analytics/analytics/src/__tests__/unit/AnalyticsListener.test.js
@@ -1,4 +1,5 @@
 /* eslint-disable */
+/* prettier-ignore */
 
 import React, { Component } from 'react';
 import { shallow, mount } from 'enzyme';