+++ design-system/badge/examples/constellation/badge-container-text-color.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import { B200, N0 } from '@atlaskit/theme/colors';
+import { B500, N0 } from '@atlaskit/theme/colors';
 
 import { Container, Format } from '../../src';
 
 export default function Example() {
   return (
-    <Container backgroundColor={B200} textColor={N0}>
+    <Container backgroundColor={B500} textColor={N0}>
       <Format>{8}</Format>
     </Container>
   );