+++ editor/editor-core/examples/88-media-embed-html-source.tsx
@@ -1,1 +1,1 @@
-export { MediaEmbedHtmlSource as default } from '@atlaskit/media-integration-test-helpers';
+export { MediaEmbedHtmlSource as default } from '@atlaskit/media-integration-test-helpers/media-embed-html-source';