+++ editor/adf-schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/adf-schema
 
+## 15.1.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 15.1.0
 
 ### Minor Changes