+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @atlaskit/editor-mobile-bridge
 
+## 25.5.3
+
+### Patch Changes
+
+- Updated dependencies
+
+## 25.5.2
+
+### Patch Changes
+
+- Updated dependencies
+
+## 25.5.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 25.5.0
 
 ### Minor Changes